ML_Project/requirements.txt
fastapi
uvicorn
pandas
torch
torchvision
pillow
matplotlib
seaborn
